get poster from IGDB
