+++ web_interface/static/favicon.ico
