+++ web_interface/templates/game.html
@@ -18,7 +18,7 @@
     </div>
     
     <div class="box">
-      <div><img src="{{ image_url }}" alt="{{ game_title }} Poster"></div>
+      <div><img src="https://{{ image_url }}" alt="{{ game_title }} Poster"></div>
       <div>
         <ul style="list-style-type:none">
           <li>Detail about Game</li>
